ruby library and config for starting apache proxy
This dynamically writes out an apache config file.  And then uses daemon_controller to start up a local copy of apache.  This can be used as an replacement for customizing your host apache config.  Or it can be used on hudson to proxy all of the services needed for running tests.