json file for designating visible labels
This file used to create a parallel json structure for adding an "on" property to an allele for display in the meiosis view. This is intended to be merged with current json data for a drake genome. Will need to change if we modify the drake genome. Perhaps this should be merged in at the GWT level.