hidden/visible changed and offset hovering labels
Originally the visible chromosome in a pair was created first placing the hidden one on top and hindering the ability to drag or show labels by hovering. The order of hidden vs. visible was changed to fix that.

Labels are now only seen when hovering over a chromosome. They are also offset from the chromosome to the left or right depending on the position of the chromosome within the container.