Add new Statechart, and move login logic to there.
The login/logout code in the statechart still delegates much of
the portal logic to the loginController, as before. This can be
changes as needed in the future.

The statechart also handles the home and activity routes. routes.js
will call a different method in statechart depending on whether the
user is logged in or out.