Add blogButton state, refactor states to make inActivity/HomePage non-concurrent
We want blogButton to be a state in loggedIn which can be concurrent with
whatever location state the user is in (inHomePage/inActivity). Unfortunately,
we can't have a mix of concurrent and non-concurrent states at the same level.

The solution is to make inHomePage and inActivity non-concurrent substates
of a new state, inLocation, and make inLocation concurrent with blogButton.

Last caveat: inLocation cannot handle state transitions between inHomePage and
inLocation, because of pivot state errors. Moved the handling of those actions 
to the deeper states.