cumulative stats
This adds cumulative statistics for a single trait. New controller now does most of the calculating for the view. Note: Some CSS styles were applied in this code, but should probably be changed in the stylesheet. This was done to get the cumulative stats to fit on the same row as the current stats.