new look for buttons
This was applied to the "bring it on" buttons as well as those on the Blog, Instructions, and Help panels. Note: The reason the hover was implemented with a background shift rather than swapping images was due to a significant delay when loading the hover image for the first time. This should probably also be applied to the "switch sex" button.