update meiosis templates to allow setting of swap button
animationView now takes a parameter called swapping the tells the view to show (true) or hide (false) the button that makes the animation jump to the frame in the animation where swapping can occur. Right now swapping is completely turned off. In the future this property can also turn on the swapping mechanism.