fixed stats view to include intermittent traits
If there was a trait that would only sometimes appear then the cumulative stats for that trait would appear only if that trait was in the current clutch. Now any traits that have accumulated stats in the cumulative stats counters will appear regardless of that characteristic being present in the current clutch.