make panels work with applets
Applets appear on top of all sproutcore views. However, an iframe will go on top of an applet, so all the movable panels were modified to have an iframe with a layout bound to the layout of the panel, but under the panel. This makes the panel appear on top of the applet. However, if two iframes are on top of each other and over the applet, then the overlapping portion of the applet comes through. Better than nothing, but not a complete solution.