fixed (hack) for set DNA not working
In order to set the DNA strand on the MW applet you need to reset the applet first and wait until it finishes reloading the model file before setting the DNA content of the model. This institutes a one second delay between the reset command and the set DNA command. A better solution would be to fix the "transcribe" button to re-enable itself when set DNA is called, or to have access to some internal value of the model such that we can check the model has finished loading before executing the set DNA.