fix iframe removal problem
To make panels appear on top of applets an iframe was added below each panel using appendChild to the main activity view. The way they were being removed caused errors when navigating to new activity pages. This is now fixed be referring to the parentView for removal instead of the current activity page which changes when navigating.