Stop loading parent (challenge) drakes, at least for now.
Loading parent drakes is too ambiguous. What if bad drakes were saved
to the back end? Right now, if the number of drakes loaded is wrong, we can
catch that. But what if the number is right? In theory, we could compare to
the authored alleles and catch that as well. But in any case, we would be
throwing away drakes and starting afresh. So some students might be getting
their old parents and some may be generating new ones. Seems better to
always generate new ones...