Use non-backend "session" property to only use current session drakes.
Challenge pool was getting messed up due to parent drakes piggybacking
by reference on saved stable drakes and being created by sproutcore.
Since they matched the challenge query, they would be loaded into the
pool. Using sessions, we can always be sure that the only drakes in the
challenge and match pools are that ones created that session.

This also makes it so the drakes are guaranteed to change when a user
switches or changes group number.