Use activity route as key for stars. === BREAKS OLD LEARNER DATA! ===
Update caselog to use only case json file, not to download activities.

We now key stars off of activity route instead of activity id, since
we'd have to ask the back-end for the activity id, and we don't want
to download all activities. Also stars will stay with user if new activity
uses an old route, and will go away if an author changes an activity 
record's route, which seems like better behavior.