Made labels look better
In some cases the frame was not large enough and the line connecting the label
to a point on the chromosome looked different if the label was on the left or right.
Adjustments were made in drawing the pointer and a monospace font was used
for the labels so it would be easier to calculate a valid frame width.