Add ability to show QTip tooltips on elements.
In the inActivity state, we can call the showInfoDialog
to append a speechbubble-like tooltip on an element.

The tooltip appears immediately, and disappears when the
element is moused out of or the bubble is clicked. From then
on the tooltip acts as an on-hover tooltip.

Add the jQuery QTip library:

 * Copyright (c) 2009 Craig Thompson
 * http://craigsworks.com
 *
 * Licensed under MIT
 * http://www.opensource.org/licenses/mit-license.php