Update biologica.js with new GenivilleDrake species.
This change also brings in the endCellInfo work that makes
it easier to determine what happened in meiosis.

GenivilleDrake species is identical to Drake species,
with the exception that the armor gene has been simplified
to just A/a.

Note: this branch *will not work* without the next commit,
which updates the application to use the new species.