Move reporting of success to endChallenge for all challenges
Most challenge types were reporting that a challenge was complete
for every successfully-matched dragon, which in a challenge where
multiple dragons needed to be matched was incorrect.

Also, this centralizes the call for the majority of the challenges to
the challenge.js version of endChallenge, which will reduce code
duplication.