Submit to /challengeProgress when challenge isn't successful
Users stay in the challenge when a match is incorrect, or when they
submit something incorrect. They're not completing the challenge,
so, since we're adding a new /challengeProgress route anyway,
it makes sense to see this is "progress" data.

More info will be added to this progress data in next commit.

NB: This makes the "win" parameter redundant, but it's left in to
help backwards compatibility.