Reset chromosome-pulldown views on activity navigation
dragonGenomeController was not getting reset alongside all
the other controllers, so chromosome views could be already
populated when navigating to them.

This was usually obscured by the fact that new challenge dragons
would be created instantly, but when that would happen slowly
(especially when LOAD_CHALLENGE_DRAGONS=true), we would see
expired dragons in the pulldown views.

[#55481718] [#54748240]