Return empty array when querying DB for nonexistent user
Currently when a new user logs in for the first time, we query
the DB with /users.json?username=nonexistent_username. This then
returns all users, which in the production db is a very large list
(~9MB...). This can take over a minute to be received by the client.

This change quickly returns the empty array instead.

The Geniverse app seems to handle this change just fine.

[#54368242]