Save dragons to backend on button push; stable saving
Added the saving functions to main, for the heck of it.

Button under breeding view should show when there are drakes in the stable
and when the record link is not visible. The record link should automatically
become visible once the breeding pen drakes are replaced by rails-backed
versions, with no further action needed on our end.