Only save bred dragons to the stable
Fixes an error with new pushing-to-backend code, where, if parents were
dragged to the stable, they would be removed from the parents pool.

There doesn't seem to be any reason to allow non-bred drakes to be
saved to the stable any more (i.e. there are no templates that
include a stable that are not breeding templates).